use int64_t for Sample

+++ src/typedefs.h
@@ -4,7 +4,7 @@
 #include <QtGlobal>
 
 typedef uint_fast32_t nFrame;
-typedef int32_t IOSample;
+typedef int64_t IOSample;
 typedef float Sample;
 
 #endif // TYPEDEFS_H
